pyproject.toml
[project]
name = "macro-econ"
version = "0.1.0"
description = "Economic Statistics Viewer - PCE, CPI, GDP, Employment"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fredapi>=0.5.2",
    "requests>=2.31",
    "pandas>=2.1",
    "pyarrow>=14.0",
    "statsmodels>=0.14",
    "scipy>=1.11",
    "numpy>=1.26",
    "plotly>=5.18",
    "jupyterlab>=4.0",
    "ipywidgets>=8.1",
    "python-dotenv>=1.0",
]

[dependency-groups]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.1",
    "mypy>=1.7",
    "pandas-stubs>=2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/macro_econ"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true

[tool.pytest.ini_options]
testpaths = ["tests"]
